#926: Split interfaces into own file
