+++ CourageScores/ClientApp/src/components/tournaments/competition.ts
@@ -4,35 +4,7 @@ import {GameMatchOptionDto} from "../../interfaces/models/dtos/Game/GameMatchOpt
 import {TournamentSideDto} from "../../interfaces/models/dtos/Game/TournamentSideDto";
 import {any} from "../../helpers/collections";
 import {repeat} from "../../helpers/projection";
-
-export interface ILayoutDataForSide {
-    id: string;
-    name: string;
-    link: JSX.Element;
-    mnemonic?: string;
-}
-
-export interface ILayoutDataForMatch {
-    sideA: ILayoutDataForSide;
-    sideB: ILayoutDataForSide;
-    scoreA: string;
-    scoreB: string;
-    bye?: boolean;
-    winner?: string;
-    saygId?: string;
-    mnemonic?: string;
-    hideMnemonic?: boolean;
-    matchOptions?: GameMatchOptionDto;
-    match?: TournamentMatchDto;
-}
-
-export interface ILayoutDataForRound {
-    name: string;
-    matches: ILayoutDataForMatch[];
-    possibleSides: TournamentSideDto[];
-    alreadySelectedSides: TournamentSideDto[];
-    round?: TournamentRoundDto;
-}
+import {ILayoutDataForMatch, ILayoutDataForRound, ILayoutDataForSide} from "./layout";
 
 export interface ITournamentLayoutGenerationContext {
     matchOptionDefaults: GameMatchOptionDto;
